% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_predict.R
\name{act_predict}
\alias{act_predict}
\title{Apply management actions as Netica Findings}
\usage{
act_predict(action_set, nodes)
}
\arguments{
\item{nodes}{}
}
\description{
helper function for simulatin grassland condition under management
}
