% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sensitivity.R
\name{analyse_sensitivity}
\alias{analyse_sensitivity}
\title{Run sensitivity analysis for a set of target nodes in a given timeslice}
\usage{
analyse_sensitivity(network, year = c(1:5))
}
\arguments{
\item{network}{an active /code{NeticaBN} object}

\item{year}{The relevant time slice for which the sensitivity analysis will be run}
}
\value{
df_mutual_info A /code{tibble} containing the variables: node_name /code{<chr>}, mutual_info /code{<dbl>}, nodelist_node_name /code{<chr>}
}
\description{
Run sensitivity analysis for a set of target nodes in a given timeslice
}

