% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_act_predict.R
\name{apply_act_predict}
\alias{apply_act_predict}
\title{Apply act_prediction function}
\usage{
apply_act_predict(strategies_df)
}
\arguments{
\item{strategies_df}{A dataframe of strategies}
}
\value{
strategies_df A dataframe with the outcoems of simulated management at the end of the specified time horizon
}
\description{
helper function for simulating grassland condition under management
}
\details{
Applies act_predict and initialise_condition functions and records the outputs in a dataframe
}

