% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_raw_dataset.R
\name{build_raw_dataset}
\alias{build_raw_dataset}
\title{Build raw field data set complete with species type and management history data}
\usage{
build_raw_dataset(raw_field_data_path, species_lookup_data_path,
  management_lookup_data_path)
}
\arguments{
\item{raw_field_data_path}{Path to the CSV file containing the raw field data, see \code{\link{lookup_species}} and \code{\link{lookup_management}}.}

\item{species_lookup_data_path}{Path to the CSV file containing the species lookup data see \code{\link{lookup_species}}.}

\item{management_lookup_data_path}{Path to the CSV file containing the management lookup data, see \code{\link{lookup_management}}.}
}
\value{
field_data_by_quadrat This is the complete field dataset in raw form containing both management history for each transect, as well as the type code for each species (i.e. row in the dataset).
}
\description{
Build raw field data set complete with species type and management history data
}
\details{
This is an internal function for building the paper vignette with the Remake package
}
