% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutual_info.R
\name{compute_mutual_info}
\alias{compute_mutual_info}
\title{Compute Mutual Information}
\usage{
compute_mutual_info(target_node, network, year)
}
\arguments{
\item{target_node}{An active /code{NeticaNode} object that is the target of inference (i.e., we want to find the influence of other nodes on this node).}

\item{network}{an active /code{NeticaBN} object}

\item{year}{The relevant time slice for which the sensitivity analysis will be run}
}
\value{
mutual_info_df A /code{tibble} with the variables /code{mutual_info} and /code{nodelist_node_name}
}
\description{
Compute Mutual Information
}
\details{
Mutual Information is computed for the target node, and calculates the mutual information of vegetation measure and management nodes in the same time slice as the target node, as well nodes in preceding time slices.
}
