% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_path.R
\name{f_path}
\alias{f_path}
\title{Get package root path}
\usage{
f_path(...)
}
\arguments{
\item{...}{character string for the filepath to be concatenated to the project directory root path.}
}
\description{
Define relative file path for overcoming working directory madness
}
\details{
Please see \url{https://github.com/krlmlr/rprojroot.git} on github for details. Note that this definition of the function checks for R package scaffolding in order to find the root path (as opposed to Rstudio files, etc).
}
