% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_case_file_data_frame.R
\name{generate_case_file_data_frame}
\alias{generate_case_file_data_frame}
\title{Generate Netica Case File as a \code{data_frame}}
\usage{
generate_case_file_data_frame(field_data_by_transect, year = c(0:5))
}
\arguments{
\item{field_data_by_transect}{The raw field data with key variables summarised over all quadrats within each transect}
}
\value{
field_data_by_management_unit \code{data_frame} containing all the variables needed for updating the GrasslandBBN with a Netica Case File.
}
\description{
Generate Netica Case File as a \code{data_frame}
}

