% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_case_file_df.R
\name{generate_case_file_df}
\alias{generate_case_file_df}
\title{Generate Netica Case File as a \code{data_frame}}
\usage{
generate_case_file_df(...)
}
\arguments{
\item{field_data_by_mu}{Field data that has been summarised by management unit. Must contain the column \code{management_unit}.}
}
\value{
casefile_df A dataframe of the same dimensions as \code{field_data_by_mu}, but with the \code{management_unit} column removed and a new \code{IDnum} column.
}
\description{
Generate Netica Case File as a \code{data_frame}
}

