% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node.R
\name{get_node}
\alias{get_node}
\title{Extract active nodes using partial matches:}
\usage{
get_node(string_to_match = character, node_names = FALSE,
  network = grassland_bbn)
}
\arguments{
\item{string_to_match}{The character string to search.}

\item{node_names}{Defaults to \code{FALSE} and returns the matched active Netica node. If \code{TRUE}, returns the names of the matched nodes. Note that matching is currently case sensitive.}

\item{network}{Character string of the active target network's name}
}
\value{
Returns a character vector of the node names that match the searched string if \code{node_names = TRUE}, else returns a list of active netica
 nodes that match the searched string if \code{node_names = FALSE}.
}
\description{
Extract active nodes using partial matches:
}
\details{
Note that because this function returns a list when \code{node_names = FALSE}, this function can handle \code{string_to_match} vectors of length greater than 1, such that if you supply a multi-element vector, a list will be returned. Note that this function no longer supports partial matches at the moment due to an error with the function returning multiple concatenated matches when a partial match is found for more than one node.
}
