% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_grassland_bbn.R
\name{learn_grassland_bbn}
\alias{learn_grassland_bbn}
\title{Learn \code{grassland_bbn} with EM Learning algorithm}
\usage{
learn_grassland_bbn(path, casefile_path, first_iteration = TRUE, save_path)
}
\arguments{
\item{path}{Filepath to a grassland_bbn Netica network}

\item{casefile_path}{A character string containing the path to the Netica casefile}

\item{first_iteration}{Logical, is this the first iteration of learning by casefile? defaults to \code{TRUE}.}
}
\value{
grassland_bbn The R object pointer to the NeticaAPI network
}
\description{
Learn \code{grassland_bbn} with EM Learning algorithm
}
\details{
Note if \code{first_iteration} is set to \code{TRUE}, node experience for target nodes will be set to 1 for every state. Please see Netica or RNetica documentation for information on the EM Learning algorithm.
}

