% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_compile_grassland_bbn.R
\name{load_compile_grassland_bbn}
\alias{load_compile_grassland_bbn}
\title{Loads Netica network into R and NeticaAPI memory and compiles network}
\usage{
load_compile_grassland_bbn(path = character)
}
\arguments{
\item{path}{a character string containing the path to the Netica file, in either \code{.dne} or \code{.neta} format.}
}
\value{
grassland_bbn the R object of network in NeticaAPI memory.
}
\description{
Loads Netica network into R and NeticaAPI memory and compiles network
}

