% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_grassland_condition.R
\name{predict_grassland_condition}
\alias{predict_grassland_condition}
\title{Predict Grassland Condition for one or more sites under one or more /code{management_strategies}}
\usage{
predict_grassland_condition(network_path = character,
  strategies_casefile = character, sites_casefile = character)
}
\arguments{
\item{network_path}{a character string containing the file path to the Netica model to be used for simulation}

\item{strategies_casefile}{a character string containing the file path to the casefile containing the /code{management_strategies}. Column names must include: /code{IDnum} and at least one of /code{Management_t} through to /code{Management_t4}.}

\item{sites_casefile}{a character string containing the file path to the casefile containing monitoring data for one or more grassland /code{sites}. Column names should include: /code{IDnum}, /code{IndigSpp_transect_t}, /code{WeedCover_t}, /code{BareGround_t}, /code{WeedDiversity_t}, /code{YearsSince_t}.}
}
\value{
A nested /code{data_frame} with the variables /code{management_unit}, /code{attr_nodes}, /code{condition_nodes},  /code{condition_horizon}, /code{GrasslandCondition_t0}, where /code{condition_horizon} is equivalent to the final time_horizon to be predicted to, as given by the number of management actions in the /code{management_strategies} casefile.
}
\description{
Enters Findings from monitoring data for 1 or more sites, and records Grassland Condition at /code{t0}. The function then applies a selection of management strategies and predicts response in grassland condition into at the final time_horizon for each scenario. This function works on a single model.
}
\details{
Note that a /code{management_strategy} consists of one or more management actions defined as node states for each management node in the model. These actions are applied sequentially through time.
}
