% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_condition.R
\name{report_condition}
\alias{report_condition}
\title{Reports Grassland Condition at t0 for a casefile of sites.}
\usage{
report_condition(network_path = character, sites_casefile = character)
}
\arguments{
\item{network_path}{a character string containing the file path to the Netica model to be used for simulation}

\item{sites_casefile}{a character string containing the file path to the casefile containing monitoring data for one or more grassland /code{sites}. Column names should include: /code{IDnum}, /code{IndigSpp_transect_t}, /code{WeedCover_t}, /code{BareGround_t}, /code{WeedDiversity_t}, /code{YearsSince_t}.}
}
\value{
A data frame with the columns management unit and Grassland Condition_t0
}
\description{
Reports Grassland Condition at t0 for a casefile of sites.
}
