% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_netica_with_license.R
\name{start_netica_with_license}
\alias{start_netica_with_license}
\title{Start Netica with License}
\usage{
start_netica_with_license(LicensePath)
}
\arguments{
\item{LicensePath}{Filepath to}
}
\value{
NeticaLicenseKey Returns \code{NeticaLicenseKey} to the parent environment.
}
\description{
The return of \code{NeticaLicenseKey} is not the main outcome of the function, but is simply a precaution to guard against potential calls of \code{library(RNetica)}, whereby RNetica will be started with new license details. Calling \code{library(RNetica)} in the same environment where \code{NeticaLicenseKey} exists will result in Netica using this license.
}

