% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by_mu.R
\name{summarise_by_mu}
\alias{summarise_by_mu}
\title{Summarise field data by transect over each management unit}
\usage{
summarise_by_mu(field_data_by_transect, year = c(0:5))
}
\arguments{
\item{field_data_by_transect}{The raw field data with key variables summarised over all quadrats within each transect}

\item{year}{an integer between 0 and 5 representing the \code{GrasslandBBN} time-slice to be parameterised from the field data stored within the case file.}
}
\value{
field_data_by_management_unit \code{data_frame} containing all the variables needed for updating the GrasslandBBN with a Netica Case File.
}
\description{
Summarise field data by transect over each management unit
}
\details{
Please ensure the following rows are supplied: \code{management_unit}, \code{BG_mean}, \code{E_mean}, \code{E_diversity}, \code{NF_diversity}, \code{management}, \code{years_since}. Note that if you are providing observations for Management period t0, the column \code{management} is still required. You can fill this character variable with \code{NA}, the function will remove it as necessary since there is no \code{t0} \code{management} period.
}
