% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by_transect.R
\name{summarise_by_transect}
\alias{summarise_by_transect}
\title{Summarise by quadrat data into per transect data with key variables for updating the GrasslandBBN}
\usage{
summarise_by_transect(field_data_by_quadrat)
}
\arguments{
\item{field_data_by_quadrat}{A tidy dataframe grouped by transect number and quadrat with multiple entries for each quadrat under the variable \code{species}. The species column includes all species encountered within the quadrat, as well as non-living entities such as Bare Ground and Litter, and non-vascular life-forms including lichen and moss. Each \code{species} has a \code{type}, an \code{origin}, and \code{growth-form} attribute, listed under the respective variable columns. Information about each transect is also included under the variables: \code{date} of survey, \code{orientation} of transect, survey \code{assistant}, last \code{management} action, \code{burn_season} of the last fire, \code{years_since} last biomass removal, and \code{biomass_reduction_year}.}
}
\value{
field_data_by_transect A data_frame containing a row for each surveyed transect. Measured variables are equivalent to the nodes in the GrasslandBBN, and include: \code{E_pc}, the mean percent cover of exotic species per transect; \code{BG_pc} mean percent area of Bare Ground for each transect; \code{NF_diversity}, the total number of native forbs in each transect; \code{E_diversity}, the total number of exotic species across each transect.
}
\description{
Summarise by quadrat data into per transect data with key variables for updating the GrasslandBBN
}

