% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_condition.R
\name{initialise_condition}
\alias{initialise_condition}
\title{Initialise condition nodes at t0}
\usage{
initialise_condition(attribute_nodes, unit)
}
\arguments{
\item{unit}{}
}
\description{
helper function for simulating grassland condition under management
}
